requirements.txt
requests
PyPDF2
pandas
python-dotenv